refactor sources into seperate module, apply moment before stacking MTSourceWM
