+++ src/heart.py
@@ -25,8 +25,6 @@ from pyrocko.fomosto import qseis, qssp
 
 from pyrocko.model import Station
 
-from pyrocko.gf.seismosizer import outline_rect_source, Cloneable
-from pyrocko.orthodrome import ne_to_latlon
 #from pyrocko.fomosto import qseis2d
 
 
@@ -45,287 +43,6 @@ lambda_sensors = {
     'JERS': 0.23513133960784313,
     'RadarSat2': 0.055465772433
     }
-
-
-class RectangularSource(gf.DCSource, Cloneable):
-    """
-    Source for rectangular fault that unifies the necessary different source
-    objects for teleseismic and geodetic computations.
-    Reference point of the depth attribute is the top-center of the fault.
-
-    Many of the methods of the RectangularSource have been modified from
-    the HalfspaceTool from GertjanVanZwieten.
-    """
-
-    width = Float.T(help='width of the fault [m]',
-                    default=1. * km)
-    length = Float.T(help='length of the fault [m]',
-                    default=1. * km)
-    slip = Float.T(help='slip of the fault [m]',
-                    default=1.)
-    opening = Float.T(help='opening of the fault [m]',
-                    default=0.)
-
-    @property
-    def dipvector(self):
-        """
-        Get 3 dimensional dip-vector of the planar fault.
-
-        Parameters
-        ----------
-        dip : scalar, float
-            dip-angle [deg] of the fault
-        strike : scalar, float
-            strike-abgle [deg] of the fault
-
-        Returns
-        -------
-        :class:`numpy.ndarray`
-        """
-
-        return num.array(
-            [num.cos(self.dip * d2r) * num.cos(self.strike * d2r),
-             -num.cos(self.dip * d2r) * num.sin(self.strike * d2r),
-              num.sin(self.dip * d2r)])
-
-    @property
-    def strikevector(self):
-        """
-        Get 3 dimensional strike-vector of the planar fault.
-
-        Parameters
-        ----------
-        strike : scalar, float
-            strike-abgle [deg] of the fault
-
-        Returns
-        -------
-        :class:`numpy.ndarray`
-        """
-
-        return num.array(
-            [num.sin(self.strike * d2r),
-             num.cos(self.strike * d2r),
-             0.])
-
-    def center(self, width):
-        """
-        Get 3d fault center coordinates. Depth attribute is top depth!
-
-        Parameters
-        ----------
-        width : scalar, float
-            width [m] of the fault (dip-direction)
-
-        Returns
-        -------
-        :class:`numpy.ndarray` with x, y, z coordinates of the center of the
-        fault
-        """
-
-        return num.array([self.east_shift, self.north_shift, self.depth]) + \
-            0.5 * width * self.dipvector
-
-    def center2top_depth(self, center):
-        """
-        Get top depth of the fault [m] given a potential center point.
-        (Patches method needs input depth to
-        be top_depth.)
-
-        Parameters
-        ----------
-        center : scalar, float
-            coordinates [m] of the center of the fault
-
-        Returns
-        -------
-        :class:`numpy.ndarray` with x, y, z coordinates of the central
-        upper edge of the fault
-        """
-
-        return num.array([center[0], center[1], center[2]]) - \
-            0.5 * self.width * self.dipvector
-
-    def bottom_depth(self, depth):
-        """
-        Get bottom depth of the fault [m].
-        (Patches method needs input depth to be top_depth.)
-
-        Parameters
-        ----------
-        depth : scalar, float
-            depth [m] of the center of the fault
-
-        Returns
-        -------
-        :class:`numpy.ndarray` with x, y, z coordinates of the central
-        lower edge of the fault
-        """
-
-        return num.array([self.east_shift, self.north_shift, depth]) + \
-            0.5 * self.width * self.dipvector
-
-    def trace_center(self, depth):
-        """
-        Get trace central coordinates of the fault [m] at the surface of the
-        halfspace.
-
-        Parameters
-        ----------
-        depth : scalar, float
-            depth [m] of the center of the fault
-
-        Returns
-        -------
-        :class:`numpy.ndarray` with x, y, z coordinates of the central
-        lower edge of the fault
-        """
-
-        bd = self.bottom_depth(depth)
-        xtrace = bd[0] - \
-            (bd[2] * num.cos(d2r * self.strike) / num.tan(d2r * self.dip))
-        ytrace = bd[1] + \
-            (bd[2] * num.sin(d2r * self.strike) / num.tan(d2r * self.dip))
-        return num.array([xtrace, ytrace, 0.])
-
-    def patches(self, nl, nw, datatype):
-        """
-        Cut source into n by m sub-faults and return n times m
-        :class:`RectangularSource` Objects.
-        Discretization starts at shallow depth going row-wise deeper.
-        REQUIRES: self.depth to be TOP DEPTH!!!
-
-        Parameters
-        ----------
-        nl : int
-            number of patches in length direction (strike)
-        nw : int
-            number of patches in width direction (dip)
-        datatype : string
-            'geodetic' or 'seismic' determines the source to be returned
-
-        Returns
-        -------
-        :class:`pscmp.PsCmpRectangularSource` or
-        :class:`pyrocko.gf.seismosizer.RectangularSource` depending on
-        datatype. Depth is being updated from top_depth to center_depth.
-        """
-
-        length = self.length / float(nl)
-        width = self.width / float(nw)
-
-        patches = []
-        for j in range(nw):
-            for i in range(nl):
-                sub_center = self.center(self.width) + \
-                    self.strikevector * ((i + 0.5 - 0.5 * nl) * length) + \
-                    self.dipvector * ((j + 0.5 - 0.5 * nw) * width)
-
-                patch = gf.RectangularSource(
-                    lat=float(self.lat),
-                    lon=float(self.lon),
-                    east_shift=float(sub_center[0]),
-                    north_shift=float(sub_center[1]),
-                    depth=float(sub_center[2]),
-                    strike=self.strike, dip=self.dip, rake=self.rake,
-                    length=length, width=width, stf=self.stf,
-                    time=self.time, slip=self.slip, anchor='center')
-
-                if nw == 1 and nl == 1:
-                    logger.warn(
-                        'RectangularSource for fault-geometry inversion'
-                        ' decimated!')
-                    if datatype == 'seismic':
-                        patch.decimation_factor = 20
-
-                    elif datatype == 'geodetic':
-                        patch.decimation_factor = 7
-
-                else:
-                    raise TypeError(
-                        "Datatype not supported either: 'seismic/geodetic'")
-
-                patches.append(patch)
-
-        return patches
-
-    def outline(self, cs='xyz'):
-        points = outline_rect_source(self.strike, self.dip, self.length,
-                                     self.width)
-        center = self.center(self.width)
-        points[:, 0] += center[0]
-        points[:, 1] += center[1]
-        points[:, 2] += center[2]
-        if cs == 'xyz':
-            return points
-        elif cs == 'xy':
-            return points[:, :2]
-        elif cs in ('latlon', 'lonlat'):
-            latlon = ne_to_latlon(
-                self.lat, self.lon, points[:, 0], points[:, 1])
-
-            latlon = num.array(latlon).T
-            if cs == 'latlon':
-                return latlon
-            else:
-                return latlon[:, ::-1]
-
-    def extent_source(self, extension_width, extension_length,
-                     patch_width, patch_length):
-        """
-        Extend fault into all directions. Rounds dimensions to have no
-        half-patches.
-
-        Parameters
-        ----------
-        extension_width : float
-            factor to extend source in width (dip-direction)
-        extension_length : float
-            factor extend source in length (strike-direction)
-        patch_width : float
-            Width [m] of subpatch in dip-direction
-        patch_length : float
-            Length [m] of subpatch in strike-direction
-
-        Returns
-        -------
-        dict with list of :class:`pscmp.PsCmpRectangularSource` or
-            :class:`pyrocko.gf.seismosizer.RectangularSource`
-        """
-        s = copy.deepcopy(self)
-
-        l = self.length
-        w = self.width
-
-        new_length = num.ceil((l + (2. * l * extension_length)) / km) * km
-        new_width = num.ceil((w + (2. * w * extension_length)) / km) * km
-
-        npl = int(num.ceil(new_length / patch_length))
-        npw = int(num.ceil(new_width / patch_width))
-
-        new_length = float(npl * patch_length)
-        new_width = float(npw * patch_width)
-        logger.info(
-            'Fault extended to length=%f, width=%f!' % (new_length, new_width))
-
-        orig_center = s.center(s.width)
-        s.update(length=new_length, width=new_width)
-
-        top_center = s.center2top_depth(orig_center)
-
-        if top_center[2] < 0.:
-            logger.info('Fault would intersect surface!'
-                        ' Setting top center to 0.!')
-            trace_center = s.trace_center(s.depth)
-            s.update(east_shift=float(trace_center[0]),
-                     north_shift=float(trace_center[1]),
-                     depth=float(trace_center[2]))
-        else:
-            s.update(east_shift=float(top_center[0]),
-                     north_shift=float(top_center[1]),
-                     depth=float(top_center[2]))
-
-        return s
 
 
 def log_determinant(A, inverse=False):
@@ -2102,7 +1819,7 @@ def discretize_sources(
 
     Parameters
     ----------
-    sources : :class:`RectangularSource`
+    sources : :class:`sources.RectangularSource`
         Reference plane, which is being extended and
     extension_width : float
         factor to extend source in width (dip-direction)
